requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
asgiref==3.3.4; python_version >= '3.6'
bleach==3.3.0
django-appconf==1.0.4
django-ckeditor==6.0.0
django-countries==7.1
django-cprofile-middleware==1.0.5
django-crispy-forms==1.11.2
django-crum==0.7.9
django-csp==3.7
django-imagekit==4.0.2
django-js-asset==1.2.2
django-js-reverse==0.9.1
django-leaflet==0.27.1
django-multiselectfield==0.1.12
django-ranged-response==0.2.0
django-rest-captcha==0.2.0
django-select2==7.7.0
django-simple-captcha==0.5.12
django-widget-tweaks==1.4.8
django==3.2
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
gunicorn==20.1.0
numpy==1.20.2; python_version >= '3.7'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pilkit==2.0
pillow==8.2.0
psycopg2==2.8.6
pyjwt==2.0.1; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-magic==0.4.22
pytz==2021.1
reverse-geocode==1.4.1
scipy==1.6.2; python_version < '3.10' and python_version >= '3.7'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
webencodings==0.5.1
